formatting and integration tests
