--- tests/integrationTests/runners/CMakeLists.txt
@@ -1,9 +1,0 @@
-target_sources(integrationTests
-        PRIVATE
-        integrationTest.cpp
-        integrationRestartTest.cpp
-
-        PUBLIC
-        integrationTest.hpp
-        integrationRestartTest.hpp
-        )
