+++ tests/regressionTests/CMakeLists.txt
@@ -3,10 +3,8 @@ target_link_libraries(regressionTests PUBLIC GTest::gtest ablateLibrary testingR
 
 target_sources(regressionTests
         PRIVATE
-        tests.cpp
+        main.cpp
         )
-
-add_subdirectory(runners)
 
 gtest_discover_tests(regressionTests
         # set a working directory so your project root so that you can find test data via paths relative to the project root
@@ -16,5 +14,4 @@ gtest_discover_tests(regressionTests
         DISCOVERY_TIMEOUT 360
         )
 
-copy_to_test_directory(${CMAKE_CURRENT_SOURCE_DIR}/outputs/)
 copy_to_test_directory(${CMAKE_CURRENT_SOURCE_DIR}/inputs/)
