--- tests/regressionTests/runners/CMakeLists.txt
@@ -1,6 +1,0 @@
-target_sources(regressionTests
-        PRIVATE
-        main.cpp
-        runners.hpp
-        runners.cpp
-        )
