--- tests/regressionTests/runners/runners.hpp
@@ -1,19 +1,0 @@
-#ifndef ABLATELIBRARY_INTEGRATIONTEST_HPP
-#define ABLATELIBRARY_RUNNERS_HPP
-
-#include <filesystem>
-#include "builder.hpp"
-#include "environment/runEnvironment.hpp"
-#include "gtest/gtest.h"
-#include "mpiTestFixture.hpp"
-#include "mpiTestParamFixture.hpp"
-#include "parameters/mapParameters.hpp"
-#include "petscsys.h"
-#include "yamlParser.hpp"
-
-/**
- * Note: the test name is assumed to be the relative path to the yaml file
- */
-class RegressionTestsSpecifier : public testingResources::MpiTestParamFixture {};
-
-#endif  // ABLATELIBRARY_INTEGRATIONTEST_HPP
