+++ tests/testingResources/asserts/stdOutAssert.hpp
@@ -1,19 +1,20 @@
 #ifndef TESTING_RESOURCE_STDOUTASSERT
 #define TESTING_RESOURCE_STDOUTASSERT
-#include "postRunAssert.hpp"
 #include "fileAssert.hpp"
+#include "postRunAssert.hpp"
 
 namespace testingResources::asserts {
 
 /**
  * Compares the std out stream to an expected log file
  */
-class StdOutAssert : public PostRunAssert, private FileAssert{
+class StdOutAssert : public PostRunAssert, private FileAssert {
    private:
     /**
      * The expected log file to compare to
      */
     const std::filesystem::path expectedLogFile;
+
    public:
     /**
      * Create an assert with the expected log file
@@ -25,10 +26,8 @@ class StdOutAssert : public PostRunAssert, private FileAssert{
      * Compares the generated output file with the expected
      * @param mpiTestFixture
      */
-     void Test(testingResources::MpiTestFixture& mpiTestFixture) override;
-
+    void Test(testingResources::MpiTestFixture& mpiTestFixture) override;
 };
-
 
 }  // namespace testingResources::asserts
