+++ tests/unitTests/domain/dmPlexTests.cpp
@@ -1,12 +1,12 @@
 #include <petsc.h>
 #include <memory>
+#include "asserts/stdOutAssert.hpp"
 #include "domain/dmPlex.hpp"
 #include "environment/runEnvironment.hpp"
 #include "gtest/gtest.h"
 #include "mpiTestFixture.hpp"
 #include "petscTestErrorChecker.hpp"
 #include "utilities/petscUtilities.hpp"
-#include "asserts/stdOutAssert.hpp"
 
 using namespace ablate;
 
@@ -44,6 +44,8 @@ TEST_P(DMPlexTestFixture, ShouldCreateAndViewDMPlex) {
 }
 
 INSTANTIATE_TEST_SUITE_P(MeshTests, DMPlexTestFixture,
-                         testing::Values((DMPlexParameters){.mpiTestParameter = testingResources::MpiTestParameter("default DMPlex", 1, "", std::make_shared<testingResources::asserts::StdOutAssert>("outputs/domain/dmPlex_NoArguments")),
-                                                            .parameters = nullptr}),
+                         testing::Values((DMPlexParameters){
+                             .mpiTestParameter = testingResources::MpiTestParameter("default DMPlex", 1, "",
+                                                                                    std::make_shared<testingResources::asserts::StdOutAssert>("outputs/domain/dmPlex_NoArguments")),
+                             .parameters = nullptr}),
                          [](const testing::TestParamInfo<DMPlexParameters> &info) { return info.param.mpiTestParameter.getTestName(); });
